backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Pydantic for schemas
pydantic==2.9.2
pydantic-settings==2.5.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2

# AWS S3 / LocalStack
boto3==1.35.36

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code quality
black==24.8.0
isort==5.13.2
flake8==7.1.1

# Utilities
email-validator==2.2.0
python-dateutil==2.9.0
openai==1.51.0
